+++ CHANGELOG.md
@@ -8,3 +8,6 @@
 - TransitionId tyoe is replaced with NodeId
 - NodeId and ContractId are now equal by value; ContractId now is NodeId wrapper
 - `ancestors()` method moved from `Transition` to `Node` trait; genesis returns an empty array
+- Consignment now has a version field
+- Anchor contains txid field
+- Consignment endpoints contains NodeId